creation/population of the evaluation package, and refactor of all classes dependant things that are now in there
