+++ src/main/java/timeseriesweka/classifiers/DD_DTW.java
@@ -4,7 +4,7 @@ import java.text.DecimalFormat;
 import utilities.ClassifierTools;
 import utilities.InstanceTools;
 import weka.classifiers.lazy.kNN;
-import utilities.ClassifierResults;
+import evaluation.ClassifierResults;
 import weka.core.DenseInstance;
 import weka.core.EuclideanDistance;
 import weka.core.Instance;
