+++ src/main/java/timeseriesweka/classifiers/ShapeletTransformClassifier.java
@@ -28,7 +28,7 @@ import timeseriesweka.classifiers.cote.HiveCoteModule;
 import timeseriesweka.classifiers.ensembles.voting.MajorityConfidence;
 import timeseriesweka.classifiers.ensembles.weightings.TrainAcc;
 import timeseriesweka.filters.shapelet_transforms.DefaultShapeletOptions;
-import utilities.ClassifierResults;
+import evaluation.ClassifierResults;
 import utilities.TrainAccuracyEstimate;
 import weka.classifiers.Classifier;
 import weka.classifiers.bayes.NaiveBayes;
