+++ src/main/java/timeseriesweka/classifiers/TSF.java
@@ -39,7 +39,7 @@ import fileIO.OutFile;
 import java.util.ArrayList;
 import java.util.Random;
 import utilities.ClassifierTools;
-import utilities.CrossValidator;
+import evaluation.CrossValidator;
 import weka.classifiers.AbstractClassifier;
 import weka.classifiers.Evaluation;
 import weka.classifiers.trees.RandomForest;
@@ -51,7 +51,7 @@ import weka.core.Instance;
 import weka.core.Instances;
 import weka.core.TechnicalInformation;
 import utilities.TrainAccuracyEstimate;
-import utilities.ClassifierResults;
+import evaluation.ClassifierResults;
 
 /*
