+++ src/main/java/timeseriesweka/classifiers/randomboss/RandomCAWPEBOSS.java
@@ -31,7 +31,7 @@ import weka.core.Capabilities;
 import weka.core.Instance;
 import weka.core.Instances;
 import weka.classifiers.Classifier;
-import utilities.ClassifierResults;
+import evaluation.ClassifierResults;
 import vector_classifiers.CAWPE;
 import weka.core.TechnicalInformation;
