+++ src/main/java/timeseriesweka/classifiers/randomboss/RandomSVMBOSS.java
@@ -29,7 +29,7 @@ import weka.core.Capabilities;
 import weka.core.Instance;
 import weka.core.Instances;
 import weka.classifiers.Classifier;
-import utilities.ClassifierResults;
+import evaluation.ClassifierResults;
 import weka.classifiers.functions.SMO;
 import weka.classifiers.functions.supportVector.Kernel;
 import weka.classifiers.functions.supportVector.PolyKernel;
