+++ src/main/java/utilities/TrainAccuracyEstimate.java
@@ -1,5 +1,6 @@
 package utilities;
 
+import evaluation.ClassifierResults;
 import weka.core.Instances;
 
 /**
