+++ src/main/java/vector_classifiers/CAWPE.java
@@ -3,7 +3,7 @@ package vector_classifiers;
 import experiments.CollateResults;
 import experiments.DataSets;
 import experiments.Experiments;
-import statistics.tests.MultipleClassifierEvaluation;
+import evaluation.MultipleClassifierEvaluation;
 import timeseriesweka.classifiers.ensembles.weightings.TrainAcc;
 import timeseriesweka.classifiers.ensembles.weightings.ModuleWeightingScheme;
 import timeseriesweka.classifiers.ensembles.weightings.TrainAccByClass;
@@ -17,7 +17,7 @@ import java.util.Arrays;
 
 import timeseriesweka.classifiers.cote.HiveCoteModule;
 import utilities.ClassifierTools;
-import utilities.CrossValidator;
+import evaluation.CrossValidator;
 import utilities.DebugPrinting;
 import utilities.InstanceTools;
 import weka.classifiers.Classifier;
@@ -36,7 +36,7 @@ import weka.filters.SimpleBatchFilter;
 import timeseriesweka.classifiers.SaveParameterInfo;
 import utilities.StatisticalUtilities;
 import utilities.TrainAccuracyEstimate;
-import utilities.ClassifierResults;
+import evaluation.ClassifierResults;
 import timeseriesweka.classifiers.ensembles.EnsembleModule;
 import timeseriesweka.classifiers.ensembles.voting.MajorityConfidence;
 import timeseriesweka.filters.SAX;
